Unify ChromaFactory
