+++ EliteChroma.sln
@@ -29,8 +29,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EliteFiles", "src\EliteFile
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EliteChroma.Core", "src\EliteChroma.Core\EliteChroma.Core.csproj", "{B52E06B6-2E15-4923-BC2D-A7A4B0A2C725}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EliteChroma.Core.Windows", "src\EliteChroma.Core.Windows\EliteChroma.Core.Windows.csproj", "{E86078D9-E03F-4A77-B64F-70A66B5ACD83}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EliteChroma", "src\EliteChroma\EliteChroma.csproj", "{D25F56A5-49FC-4B52-8465-B296ECF10CAE}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{140A3D09-123F-49F0-A523-48C46834E416}"
@@ -59,10 +57,6 @@ Global
 		{B52E06B6-2E15-4923-BC2D-A7A4B0A2C725}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B52E06B6-2E15-4923-BC2D-A7A4B0A2C725}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B52E06B6-2E15-4923-BC2D-A7A4B0A2C725}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E86078D9-E03F-4A77-B64F-70A66B5ACD83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E86078D9-E03F-4A77-B64F-70A66B5ACD83}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E86078D9-E03F-4A77-B64F-70A66B5ACD83}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E86078D9-E03F-4A77-B64F-70A66B5ACD83}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D25F56A5-49FC-4B52-8465-B296ECF10CAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D25F56A5-49FC-4B52-8465-B296ECF10CAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D25F56A5-49FC-4B52-8465-B296ECF10CAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -88,7 +82,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{76753B28-FD0F-40E5-934E-15EAC3FC113C} = {D121ADAF-389A-4906-AC13-99A8B4973C9C}
 		{B52E06B6-2E15-4923-BC2D-A7A4B0A2C725} = {D121ADAF-389A-4906-AC13-99A8B4973C9C}
-		{E86078D9-E03F-4A77-B64F-70A66B5ACD83} = {D121ADAF-389A-4906-AC13-99A8B4973C9C}
 		{D25F56A5-49FC-4B52-8465-B296ECF10CAE} = {D121ADAF-389A-4906-AC13-99A8B4973C9C}
 		{0CD88E57-527E-4EEE-B48B-83AB2E11E893} = {140A3D09-123F-49F0-A523-48C46834E416}
 		{D67C02E2-8CCC-4867-8806-A9AF3619F6C5} = {140A3D09-123F-49F0-A523-48C46834E416}
