--- src/EliteChroma.Core.Windows/.gitignore
@@ -1,3 +1,0 @@
-bin/
-obj/
-*.user
