--- src/EliteChroma.Core.Windows/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using System;
-
-[assembly: CLSCompliant(true)]
