--- src/EliteChroma.Core.Windows/EliteChroma.Core.Windows.csproj
@@ -1,32 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0-windows</TargetFramework>
-    <IsPublishable>false</IsPublishable>
-  </PropertyGroup>
-
-  <Import Project="$(MSBuildThisFileDirectory)..\..\Common.props" />
-
-  <PropertyGroup>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <NoWarn>1701;1702;1591;SA1600;SA1602</NoWarn>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <NoWarn>1701;1702;1591;SA1600;SA1602</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\EliteChroma.Core\EliteChroma.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>EliteChroma.Core.Tests</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-
-</Project>
